requirements.txt
Django>=3.2.25,<4.0
djangorestframework>=3.15.1
gunicorn>=20.0.4
